[DEV-667] Fixed migration
